Move NumberOrString up to the general API package
